Adds a success screen
