+++ app/(gcforms)/[locale]/(test)/test/components/server/Success.tsx
@@ -1,0 +1,7 @@
+export const Success = () => {
+  return (
+    <>
+      <h1>Success</h1>
+    </>
+  );
+};
